refactor: use only data.ts in crm page
